# api-docs/api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443
  description: Generated server url
paths:
...
  /api/trainings/{trainingId}:
    get:
      tags:
      - training-rest-controller
      summary: Get a training by ID
      operationId: getTrainingByIdd
      parameters:
      - name: trainingId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Training not found
          content: {}
        "200":
          description: Training found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UniqueTrainingDTO"
        "400":
          description: Bad request - Invalid parameters
        "401":
          description: Unauthorized access - Authentication is required
        "403":
          description: Forbidden - You don't have permission to access
        "500":
          description: Internal server error
    put:
      tags:
      - training-rest-controller
      summary: Replace a training
      operationId: replacePost
      parameters:
      - name: trainingId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrainingDTO"
        required: true
      responses:
        "404":
          description: Training not found
          content: {}
        "200":
          description: Training replaced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingDTO"
        "400":
          description: Bad request - Invalid parameters
        "401":
          description: Unauthorized access - Authentication is required
        "403":
          description: Forbidden - You don't have permission to access
        "500":
          description: Internal server error
    delete:
      tags:
      - training-rest-controller
      summary: Delete a training
      operationId: deleteTraining
      parameters:
      - name: trainingId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Training not found
          content: {}
        "200":
          description: Training deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingDTO"
        "400":
          description: Bad request - Invalid parameters
        "401":
          description: Unauthorized access - Authentication is required
        "403":
          description: Forbidden - You don't have permission to access
        "500":
          description: Internal server error
...
  /api/trainingComments/valid:
    put:
      tags:
      - training-comment-restcontroller
      summary: Unreport training comment
      description: Unreport training comment
      operationId: unreportComment
      parameters:
      - name: commentId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Training comment not unreported
        "200":
          description: Training comment unreported
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TrainingCommentDTO"
        "400":
          description: Bad request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TrainingCommentDTO"
        "403":
          description: Forbidden-Access denied
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TrainingCommentDTO"
  /api/trainingComments/report:
    put:
      tags:
      - training-comment-restcontroller
      summary: Report training comment
      description: Report training comment
      operationId: reportComment
      parameters:
      - name: commentId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Training comment not reported
        "200":
          description: Training comment reported
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TrainingCommentDTO"
        "400":
          description: Bad request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TrainingCommentDTO"
        "403":
          description: Forbidden-Access denied
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TrainingCommentDTO"
  /api/trainingComments/:
    get:
      tags:
      - training-comment-restcontroller
      summary: Get paginated training comments
      description: Get paginated training comments
      operationId: getPaginatedTrainingComments
      parameters:
      - name: trainingId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "404":
          description: Training comments not found
        "200":
          description: Training comments found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrainingCommentDTO"
        "400":
          description: Bad request
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrainingCommentDTO"
    put:
      tags:
      - training-comment-restcontroller
      summary: Update a training comment
      description: Update a training comment
      operationId: updateTrainingComment
      parameters:
      - name: id
        in: query
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrainingCommentDTO"
        required: true
      responses:
        "404":
          description: Training comment not updated
        "200":
          description: Training comment updated
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TrainingCommentDTO"
        "400":
          description: Bad request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TrainingCommentDTO"
        "403":
          description: Forbidden-Access denied
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TrainingCommentDTO"
    post:
      tags:
      - training-comment-restcontroller
      summary: Create a new training comment
      description: Create a new training comment
      operationId: createTrainingComment
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrainingCommentDTO"
        required: true
      responses:
        "404":
          description: Training comment not created
        "201":
          description: Training comment created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TrainingCommentDTO"
        "400":
          description: Bad request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TrainingCommentDTO"
        "403":
          description: Forbidden-Access denied
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TrainingCommentDTO"
    delete:
      tags:
      - training-comment-restcontroller
      summary: Delete a training comment
      description: Delete a training comment
      operationId: deleteTrainingComment
      parameters:
      - name: id
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Training comment not deleted
        "200":
          description: Training comment deleted
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TrainingCommentDTO"
        "400":
          description: Bad request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TrainingCommentDTO"
        "403":
          description: Forbidden-Access denied
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TrainingCommentDTO"
...
  /api/nutritionComments/valid:
    put:
      tags:
      - nutrition-comment-rest-controller
      summary: Unreport nutrition comment
      description: Unreport nutrition comment
      operationId: unreportComment_1
      parameters:
      - name: commentId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Nutrition comment not found
        "200":
          description: Nutrition comment unreported
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NutritionCommentDTO"
        "400":
          description: Bad request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NutritionCommentDTO"
        "403":
          description: Forbidden-Access denied
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NutritionCommentDTO"
  /api/nutritionComments/report:
    put:
      tags:
      - nutrition-comment-rest-controller
      summary: Report nutrition comment
      description: Report nutrition comment
      operationId: reportComment_1
      parameters:
      - name: commentId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Nutrition comment not found
        "200":
          description: Nutrition comment reported
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NutritionCommentDTO"
        "400":
          description: Bad request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NutritionCommentDTO"
        "403":
          description: Forbidden-Access denied
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NutritionCommentDTO"
...
  /api/trainings/:
    get:
      tags:
      - training-rest-controller
      summary: Get all trainings
      operationId: getTrainings
      responses:
        "404":
          description: Trainings not found
          content: {}
        "200":
          description: Found all trainings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingDTO"
        "400":
          description: Bad request - Invalid parameters
        "401":
          description: Unauthorized access - Authentication is required
        "403":
          description: Forbidden - You don't have permission to access
        "500":
          description: Internal server error
    post:
      tags:
      - training-rest-controller
      summary: Create a training
      operationId: createTraining
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrainingDTO"
        required: true
      responses:
        "404":
          description: Not Found
        "201":
          description: Training created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingDTO"
        "400":
          description: Bad request - Invalid parameters
        "401":
          description: Unauthorized access - Authentication is required
        "403":
          description: Forbidden - You don't have permission to access
        "500":
          description: Internal server error
...
  /api/trainingComments/{trainingId}/:
    get:
      tags:
      - training-comment-restcontroller
      summary: Get training comments by training id
      description: Get training comments by training id
      operationId: getTrainingCommentsByTrainingId
      parameters:
      - name: trainingId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Training comments not found
        "200":
          description: Training comments found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrainingCommentDTO"
        "400":
          description: Bad request
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrainingCommentDTO"
  /api/trainingComments/all:
    get:
      tags:
      - training-comment-restcontroller
      summary: Get all training comments
      description: Get all training comments
      operationId: getAllTrainingComments
      responses:
        "404":
          description: Training comments not found
        "200":
          description: Training comments found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrainingCommentDTO"
        "400":
          description: Bad request
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrainingCommentDTO"
...
components:
  schemas:
    UserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        roles:
          type: array
          items:
            type: string
        email:
          type: string
...
    TrainingDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        goal:
          type: string
        intensity:
          type: string
    TrainingCommentDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        isNotified:
          type: boolean
        training:
          $ref: "#/components/schemas/TrainingDTO"
        user:
          $ref: "#/components/schemas/UserDTO"
    NutritionDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        image:
          type: boolean
        goal:
          type: string
        calories:
          type: integer
          format: int32
    NutritionCommentDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
        name:
          type: string
        isNotified:
          type: boolean
        nutrition:
          $ref: "#/components/schemas/NutritionDTO"
        user:
          $ref: "#/components/schemas/UserDTO"
...
    UniqueTrainingDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        goal:
          type: string
        intensity:
          type: string